moves HireMe content to its own folder, adds router to App
