+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/trevorhewitt/Documents/Code/mysite/src/index.js":"1","/Users/trevorhewitt/Documents/Code/mysite/src/Background.tsx":"2","/Users/trevorhewitt/Documents/Code/mysite/src/utility.ts":"3","/Users/trevorhewitt/Documents/Code/mysite/src/Links.tsx":"4"},{"size":219,"mtime":1604697428788,"results":"5","hashOfConfig":"6"},{"size":4982,"mtime":1604697111547,"results":"7","hashOfConfig":"6"},{"size":1263,"mtime":1604697119647,"results":"8","hashOfConfig":"6"},{"size":1755,"mtime":1606360623101,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1ie4cm2",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/trevorhewitt/Documents/Code/mysite/src/index.js",[],"/Users/trevorhewitt/Documents/Code/mysite/src/Background.tsx",[],"/Users/trevorhewitt/Documents/Code/mysite/src/utility.ts",[],"/Users/trevorhewitt/Documents/Code/mysite/src/Links.tsx",[]]
+[{"/Users/trevorhewitt/Documents/Code/mysite/src/index.js":"1","/Users/trevorhewitt/Documents/Code/mysite/src/utility.ts":"2","/Users/trevorhewitt/Documents/Code/mysite/src/HireMe/Background.tsx":"3","/Users/trevorhewitt/Documents/Code/mysite/src/App.tsx":"4","/Users/trevorhewitt/Documents/Code/mysite/src/HireMe/ReferenceLinks.tsx":"5","/Users/trevorhewitt/Documents/Code/mysite/src/HireMe/AboutMe.tsx":"6","/Users/trevorhewitt/Documents/Code/mysite/src/HireMe/index.tsx":"7"},{"size":219,"mtime":1604697428788,"results":"8","hashOfConfig":"9"},{"size":1263,"mtime":1604697119647,"results":"10","hashOfConfig":"9"},{"size":4983,"mtime":1606370371030,"results":"11","hashOfConfig":"9"},{"size":362,"mtime":1606531854270,"results":"12","hashOfConfig":"9"},{"size":1740,"mtime":1606370400891,"results":"13","hashOfConfig":"9"},{"size":824,"mtime":1606531149689,"results":"14","hashOfConfig":"9"},{"size":977,"mtime":1606531500106,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1ie4cm2",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/trevorhewitt/Documents/Code/mysite/src/index.js",[],"/Users/trevorhewitt/Documents/Code/mysite/src/utility.ts",[],"/Users/trevorhewitt/Documents/Code/mysite/src/HireMe/Background.tsx",[],"/Users/trevorhewitt/Documents/Code/mysite/src/App.tsx",[],"/Users/trevorhewitt/Documents/Code/mysite/src/HireMe/ReferenceLinks.tsx",["30"],"/Users/trevorhewitt/Documents/Code/mysite/src/HireMe/AboutMe.tsx",[],"/Users/trevorhewitt/Documents/Code/mysite/src/HireMe/index.tsx",[],{"ruleId":"31","severity":1,"message":"32","line":46,"column":7,"nodeType":"33","messageId":"34","endLine":46,"endColumn":27},"@typescript-eslint/no-redeclare","'Link' is already defined.","Identifier","redeclared"]
